mcp_server/pyproject.toml
[project]
name = "file-reader-mcp"
version = "0.2.0"
description = "MCP file reader server supporting multiple document formats"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "pandas>=2.0.0",
    "python-docx>=1.1.0",
    "PyMuPDF>=1.23.0",
]

[project.scripts]
file-reader-mcp = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
]